+++ Changes.md
@@ -1,6 +1,8 @@
 ## WIP  TBD
 
- * Corrected documentation for deferred.Error which gave incorrect guidance on how to use it.
+ * Adding `bytes.ContainsOnly`, `bytes.FromRange`, `bytes.Reverse`, `bytes.Indent`.
+ * Adding `strings.Indent`.
+ * Corrected documentation for `deferred.Error` which gave incorrect guidance on how to use it.
 
 ## v0.7.0  2024-06-13